Create conditional.c
This programe is  base on conditional statement. 
Conditional statement are if,else,else if,nested.
In this program we use if and else if ststement.
when we run the program first we give the input.
They ask "Enter The Marks" then we can enter any integer number. When we enter integer the further program is run.
When we enter "90" they print "Grade A".